+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 v4.0.7 - stable
- * simpler account selector - permanent indicator on the left
- * simplify: remove most-recent-used account in hamb menu
+ * simpler account selector - permanent indicator to the left
+ * simplify: remove most-recent-used in hamb menu
+ 
+ v4.0.9 - dev
+ * hide network selection while selecting accounts
+ * position approve popup on the right
+